Use cosine distance
This is waaaay better at finding ULAN matches than the stringdist default that considers edit distance.

I've temporariliy scrapped the "score" scaler - this was silly. Ideally it would be good to have the "local" and "sparql" scores work along roughly the same scale, but I will need to do more testing to really figure that out.