Construct query table with multiple SPARQL calls
The SPARQL call for the complete table is way too large, and requires downloading a lot of duplicate information. As an alternative, we can download just the alntames, and a table of the attributes, and do the joining within R. This is saved as an .rds file in the ulanrdb package installation.